Refactor internal logger, remove SDK error logging to sentry
